fix(viewer): address PR #723 review feedback
CodeRabbit + Codex flagged a batch of issues from the rotation /
wall-endpoint commit. Addressing the small/clear ones now and
documenting two larger items for follow-up.

Fixed:
- readEntityRotation / readWallEndpoints / GeometryEditCard /
  GizmoOverlay now go through getOrCreateStoreEditor (or the new
  readEntityPosition slice action). Move + rotate + endpoint
  handles surface on first selection of a freshly-loaded model
  instead of waiting for an unrelated mutation to prime the
  editor cache.
- Cesium setCesiumPlacementEditMode(false) now clears the
  placement draft state, matching the toggle's disable branch.
- Viewport.unprojectToFloor + selectionHandlers.raycastStoreyFloor
  scale the cursor to drawing-buffer coordinates before calling
  camera.unprojectToRay. Fixes pick drift at DPR ≠ 1 where
  endpoint handles would land off-cursor.
- GizmoOverlay + WallEndpointOverlay RAF signature now includes
  projectionMode + orthoSize so ortho-zoom-only changes wake the
  overlay.
- GizmoOverlay only advances its per-axis accumulator when
  translateEntity succeeds, keeping drag state and model state in
  sync if the mutation rejects mid-drag.
- rotateEntity refuses entities with an implicit RefDirection
  rather than materialising an IfcDirection that would be
  orphaned on undo. Every in-store builder (addColumn / addWall /
  …) emits an explicit RefDirection, so the refusal only trips on
  hand-rolled source-buffer entities — better than leaking
  entities through the export.
- placement-edit.ts split (per AGENTS.md §7 ~400-line rule) into:
    placement-core.ts (340 LOC) — generic chain + product helpers
    wall-edit.ts      (193 LOC) — rectangle-wall resize
    placement-edit.ts ( 42 LOC) — barrel re-export
- docs/design/level-display-modes.md fenced block gets a `text`
  language tag for MD040.

Deferred (with rationale):
- Codex P2: mesh re-render on translate / rotate / resize.
  Architecturally significant — needs a renderer-side per-entity
  mesh-position-update path (or a translation channel parallel to
  the existing color-overrides). Will plan separately.
- CodeRabbit major: resizeWall atomic undo. Currently four
  positional mutations per drag frame land as four undo entries.
  Needs a batched-mutation primitive on the store
  (setPositionalAttributesBatch); planned as a follow-up.
